requirements.txt
anthropic
fastapi
python-dotenv
python-multipart
pymongo
uvicorn